REF: renames modal related enities
