--- GetF!t/ModalType.swift
@@ -1,11 +1,0 @@
-import Foundation
-
-enum ModalType {
-    case timer
-    case history
-    case successView
-}
-
-extension ModalType: Identifiable {
-    var id: Self { self }
-}
